+++ ChangeLog.md
@@ -6,6 +6,7 @@ Eventum Issue Tracking System
 
 - Update custom fields from update issue page (Bryan Alsdorf, #88)
 - Allow time category/summary to be set when sending emails (Bryan Alsdorf)
+- Add missing 'Scheduled Release' field back to update page (Bryan Alsdorf, #89)
 
 2015-11-02, Version 3.0.5
 -------------------------